refactor: extract runtime html content concern
